Break the db tangle's spine by homing the reachability core
scopeIsChildDeadEnd and its scopeReachesPage helper are the reachability
core shared by the edge writer and the modifier writer — their own doc
says the two blocks must never diverge. That core lived in
modifier-resolve, so listing-relationship-validation reached into it, and
that back-edge closed a seven-module cycle through groups, membership,
listing parents, prices, and records.

The core moves to listing-parents-rules, the pure leaf both writers
already lean on, and its tests move with it. The seven-module group
dissolves: measured on load-time edges, the tree now holds no db cycle
bigger than the records-prices-parents triangle and the groups-membership
pair — the write pipeline's two genuine remaining loops.
