+++ src/shared/listing-parents-rules.ts
@@ -122,3 +122,34 @@ const EDGE_ERROR_RULES: readonly EdgeReason[] = [
  * checked separately by the editor.
  */
 export const edgeFieldError: EdgeReason = firstReason(EDGE_ERROR_RULES);
+
+/** Whether a resolved listing scope is reachable from a page's listing ids:
+ * a whole-order scope (null) always is; a listing set must share an id. Shared
+ * by the add-on listing and the child-reachability hard block. */
+export const scopeReachesPage = (
+  scope: number[] | null,
+  pageIds: Set<number>,
+): boolean => scope === null || scope.some((id) => pageIds.has(id));
+
+/**
+ * The single reachability test shared by both child-scoped-add-on hard blocks,
+ * so the edge save and the modifier save can never diverge.
+ *
+ * An opt-in add-on is a dead end exactly when its resolved scope is a listing
+ * set that names a suppressed child yet reaches none of the pages that would
+ * load it. A whole-order scope (`null`) is reachable everywhere.
+ *
+ * Callers define "reachable" from their own side, which is why the two id sets
+ * are parameters rather than derived here.
+ */
+export const scopeIsChildDeadEnd = (
+  scope: number[] | null,
+  suppressed: Set<number>,
+  reachable: Set<number>,
+): boolean => {
+  if (scope === null) return false;
+  return (
+    scope.some((id) => suppressed.has(id)) &&
+    !scopeReachesPage(scope, reachable)
+  );
+};
